修改了AbilityEffect的结构，新增可携带参数的Entry类型，并统一为IExecutableEffect 添加了CompositeTrigger以减少原始依赖 修改了掉落物、卡牌等，将其改为通过CompositeTrigger触发对应效果 修复了MegaBattery相关逻辑

+++ src/main/java/net/luojiuoscar/isaac_disaster/event/TriggerModuleEvents.java
@@ -48,7 +48,7 @@ public class TriggerModuleEvents {
         queue.lock();
 
         for (var inst : queue.getQueue()) {
-            context.set(ContextKeys.AMPLIFIER, inst.stacks);
+            context.set(ContextKeys.AMPLIFIER, (double) inst.stacks);
             ITriggerModule module = reg.getValue(inst.id);
             if (module == null) continue;
 
@@ -58,7 +58,7 @@ public class TriggerModuleEvents {
     }
 
     public static void bulletDispatch(AbilityEffectContext context, TriggerType type){
-        context.set(ContextKeys.AMPLIFIER, 1);
+        context.set(ContextKeys.AMPLIFIER, 1.);
         ModTriggerModule.BULLET_TRIGGER_MODULE.get().fire(context, type);
     }
